+++ fiboa_cli/datasets/de_nrw.py
@@ -4,6 +4,7 @@ URI = "https://www.opengeodata.nrw.de/produkte/umwelt_klima/bodennutzung/landwir
 ID = "de_nrw"
 TITLE = "Field boundaries for North Rhine-Westphalia (NRW), Germany"
 DESCRIPTION = """A field block (German: "Feldblock") is a contiguous agricultural area surrounded by permanent boundaries, which is cultivated by one or more farmers with one or more crops, is fully or partially set aside or is fully or partially taken out of production. Field blocks are classified separately according to the main land uses of arable land, grassland, permanent crops, 2nd pillar and other. Since 2005, field blocks in NRW have represented the area reference within the framework of the Integrated Administration and Control System (IACS) for EU agricultural subsidies."""
+# From http://osmtipps.lefty1963.de/2008/10/bundeslnder.html
 BBOX = [5.8659988131,50.3226989435,9.4476584861,52.5310351488]
 EXTENSIONS = ["https://fiboa.github.io/inspire-extension/v0.1.0/schema.yaml"]
 COLUMNS = {
@@ -31,6 +32,6 @@ MISSING_SCHEMAS = {
 
 def convert(output_file, cache_file = None):
     """
-    Converts the DE NRW field boundary datasets to fiboa.
+    Converts the NRW (Germany) field boundary datasets to fiboa.
     """
     convert_(output_file, cache_file, URI, COLUMNS, ID, TITLE, DESCRIPTION, BBOX, EXTENSIONS, MISSING_SCHEMAS)